fix: move blake2b to another file
